requirements.txt
nicegui>=3.6,<4.0
pyserial>=3.5,<4.0
pyyaml>=6.0,<7.0